Fixed code after merge with develop
The code stopped working after the merge, so I did some fixes.

The changes to index.html and AppECDC.svelte are cosmetic, they are
supposed to make the code more similar to the latest code in the
repository.

For spot.tx, I reverted to a previous version of the code, because
the new version did not work. I had to tweak the UUID generation
so that it works behind a reverse proxy.