Added Refguides and corrected CareTeams for patient count
Refguides is one of the three CareTeams attributes whose population
counts are being reported.

However, there will be one CareTeams instance for every patient in
the FHIR store and we are only interested in the number of Laboratories
or whatever associated with the found patients. So in
ResultTableComponent.wc.svelte, divide by the patient count for the
relevant columns.