Use stratifiers to populate Labs, Hospitals and Refguides
Previously, I thought that I could (and maybe should) use an appropriate
data model plus CQL to find counts for laboratories, etc. But I never
found a way to get it to work, I always fell foul of the fact that
I would get the sum of counts for all patients.

One morning I woke up with the idea in my mind that I could get this
information by counting the number of non-zero bins in the relevant
stratifier. I tried it and it worked, hence this commit.

I have removed the stuff that I previously implemented for this
functionality.