chore: stricter catalogue typing and more JSDoc
This commit makes the typing of the catalogue stricter. Specifically it turns the Category type into a tagged union and adds individual variants for the different types of catalogue items: e.g. AutocompleteCategory and NumericRangeCategory.

This commit also gets rid of the TreeNode type which required a few additional changes in "stores/catalogue.ts". (Everything I touched was marked for deprecation before anyways and will be removed in a future version).

This commit also renames treeData.ts to catalogue.ts in preparation for auto-generating JSON schema from TypeScript.