chore!: use ajv for JSON schema validation instead of @exodus/schemasafe
The advantage of ajv is that is provides a CLI and it also seems to be more widely used.

This commit also replaces catalogue.schema.json with a version that is automatically generated from the catalogue.ts.

BREAKING CHANGE: The catalogue.json of your project needs to be adjusted. Category groups, i.e. objects that have the `childCategories` property now require the additional property `"fieldType": "group"`.