chore: change error toasts back to use writable()
In Svelte 5, $state() can in theory replace writable() everywhere but this has some consequences. In the case of error toasts showToast adds the toast to the current list of toasts. Adding to a list counts as a read and a write in Svelte's reactivity system. If you have a read and a write to the same reactive  variable in an $effect() this triggers an infinite update loop that can only be prevented by using untrack(). This is explained at  the very end of https://svelte.dev/docs/svelte/$effect#When-not-to-use-$effect. Knowing when to use untrack() requires all the knowledge I just wrote down and not everyone might have that. Without untrack() people get an infinite update loop which will confuse them. Therefore I'm reverting for now to using writable() again so that untrack() is not needed.